create module structure
