+++ js/id-generator-creation.js
@@ -1,0 +1,9 @@
+const createIdGenerator = () => {
+  let generatedId = 0;
+  return () => {
+    generatedId += 1;
+    return generatedId;
+  };
+};
+
+export {createIdGenerator};
